[add] added interface Utils to manage common tasks
