+++ api/src/main/java/it/unicam/cs/mpmgc/vectorrally/api/controller/match/CLIGameEngine.java
@@ -15,11 +15,11 @@ import it.unicam.cs.mpmgc.vectorrally.api.view.CLIIOController;
 import java.util.List;
 
 /**
- * This class implements the GameEngine interface, managing the game loop and overall game logic.
+ * Represents the game engine for the vector rally game using a command-line interface.
+ * Manages the game loop and overall game logic.
  *
  * @version 1.0
  * @since 2024-07-11
- *
  * @author Marta Musso
  * <a href="mailto:marta.musso@studenti.unicam.it">marta.musso@studenti.unicam.it</a>
  */
@@ -30,6 +30,11 @@ public class CLIGameEngine implements GameEngine {
     private RaceTrack raceTrack;
     private final GameSetup setup;
 
+    /**
+     * Constructs a CLIGameEngine with the specified IO controller.
+     *
+     * @param ioController the IO controller used for input/output operations
+     */
     public CLIGameEngine(CLIIOController ioController) {
         this.ioController = ioController;
         RaceTrackBuilder raceTrackBuilder = new RaceTrackBuilder();
@@ -46,10 +51,17 @@ public class CLIGameEngine implements GameEngine {
         }
     }
 
+    /**
+     * Sets up the match by initializing the track, players, and shift algorithm.
+     * Repeats the setup process until the user confirms the configuration.
+     *
+     * @return the result of the setup containing the neighbors generator, racetrack, and players
+     * @throws Exception if an error occurs during setup
+     */
     private SetupResult setupMatch() throws Exception {
         boolean confirmConfiguration = false;
         while (!confirmConfiguration) {
-            this.neighborsGenerator = this.setup.initializeShiftAlgorithm();
+            this.neighborsGenerator = this.ioController.initializeShiftAlgorithm();
             this.raceTrack = this.setup.initializeTrack();
             this.players = this.setup.initializePlayers(raceTrack);
             confirmConfiguration = ioController.askIfSatisfiedWithConfiguration(raceTrack, players);
@@ -57,12 +69,25 @@ public class CLIGameEngine implements GameEngine {
         return new SetupResult(neighborsGenerator, raceTrack, players);
     }
 
+    /**
+     * Starts the match with the given players, racetrack, and neighbors generator.
+     *
+     * @param players the list of players participating in the match
+     * @param raceTrack the racetrack on which the match is played
+     * @param neighborsGenerator the generator for neighboring moves
+     * @throws Exception if an error occurs during match initialization or execution
+     */
     private void startMatch(List<Player> players, RaceTrack raceTrack, NeighborsGenerator neighborsGenerator) throws Exception {
         MatchController matchController = new CLIMatchController(ioController, new BasicMovesGenerator<>(neighborsGenerator, new BasicMoveValidator()));
         matchController.initializeMatch(players, raceTrack);
         matchController.startMatch();
     }
 
+    /**
+     * Handles the end of the match, displaying an end message and asking the user if they want to play another match.
+     *
+     * @return true if the user wants to play another match, false otherwise
+     */
     private boolean endMatch() {
         ioController.displayEndMatchMessage();
         return ioController.askToPlayAnotherMatch();
