+++ api/src/main/java/it/unicam/cs/mpmgc/vectorrally/api/controller/match/GameEngine.java
@@ -5,14 +5,15 @@ package it.unicam.cs.mpmgc.vectorrally.api.controller.match;
  *
  * @version 1.0
  * @since 2024-07-11
- *
- * @autor Marta Musso
+ * @author Marta Musso
  * <a href="mailto:marta.musso@studenti.unicam.it">marta.musso@studenti.unicam.it</a>
  */
 public interface GameEngine {
 
     /**
      * Starts the game application.
+     *
+     * @throws Exception if an error occurs while starting the game.
      */
     void startGame() throws Exception;
 }
